start client-portfolio phone

+++ util/file.js
@@ -1,11 +1,13 @@
 const fs = require("fs");
 
 const deleteFile = (filePath) => {
-  fs.unlink(filePath, (err) => {
-    if (err) {
-      throw err;
-    }
-  });
+  if (filePath) {
+    fs.unlink(filePath, (err) => {
+      if (err) {
+        console.log(err);
+      }
+    });
+  }
 };
 
 exports.deleteFile = deleteFile;
